Add ability to determine whether window has been moved atleast once
This is useful for example: A video player, you want it to play / stop
when ACTION_UP is called, but not when you're dragging the view and then you
move your finger up. This way, "moved" will be true and you will be able to avoid
the ACTION_UP on drag, but you have access to setMoved, and you can set it to false
when you consideer.