Fix orientation change issue, added ACTION_TOGGLE_VIS
ACTION_TOGGLE_VIS and the code to handle it is for toggling visibility from the persisten notification alternative, instead of having additional notification to show the window. To disable hidden notification, slightly modified the code to ignore it, if the overwritten getHiddenNotificationIntent() returns null.

To handle orientation changes, added onConfigurationChanged() call, which calls Window.updateScreenSize().